@@ -1,8 +1,8 @@
-const users = [];
+const users: UserType[] = [];
 
-const addUser = ({ id, name, room }) => {
-  name = name.trim().toLowerCase();
-  room = room.trim().toLowerCase();
+export const addUser = ({ id, name, room }: UserType) => {
+  name = name?.trim().toLowerCase();
+  room = room?.trim().toLowerCase();
 
   const existingUser = users.find((user) => user.room === room && user.name === name);
 
@@ -16,14 +16,12 @@ const addUser = ({ id, name, room }) => {
   return { user };
 };
 
-const removeUser = (id) => {
+export const removeUser = (id: string) => {
   const index = users.findIndex((user) => user.id === id);
 
   if (index !== -1) return users.splice(index, 1)[0];
 };
 
-const getUser = (id) => users.find((user) => user.id === id);
+export const getUser = (id: string) => users.find((user) => user.id === id);
 
-const getUsersInRoom = (room) => users.filter((user) => user.room === room);
-
-module.exports = { addUser, removeUser, getUser, getUsersInRoom };
+export const getUsersInRoom = (room: string) => users.filter((user) => user.room === room);
