@@ -1,15 +1,13 @@
 import React from "react";
-
-import { BrowserRouter as Router, Route } from "react-router-dom";
-
+import { Route, Switch } from "react-router-dom";
 import Join from "./components/Join";
 import Chat from "./components/Chat";
 
-const App = () => (
-  <Router>
+const App: React.FC = () => (
+  <Switch>
     <Route path="/" exact component={Join} />
     <Route path="/chat" component={Chat} />
-  </Router>
+  </Switch>
 );
 
 export default App;
