@@ -1,13 +1,13 @@
-import React from "react";
 import { StrictMode } from "react";
 import ReactDOM from "react-dom";
 import { BrowserRouter } from "react-router-dom";
-
+import GlobalStyles from "globalStyles";
 import App from "./App";
 
 ReactDOM.render(
   <StrictMode>
     <BrowserRouter>
+      <GlobalStyles />
       <App />
     </BrowserRouter>
   </StrictMode>,
